Too Unwieldy & Not Enough Compr Testing w/ Macbeth
fz1p.c (the evolution of ziptxt.c) is consistently flawless in terms of its ability to losslessly compress files of both any size and with any text. 

Ziptxt's approach was fundamentally flawed from the start, as it had been designed to compress English really well rather than any text whatsoever. That, coupled with the fact it started choke after running files > just a mere 5K bytes-worth of memory made it too unpredictable to have my stamp of approval.

Thus I'm mothballing the code for now, as there are still some neat algorithms with respect to password-based encryption & substring identification to possibly dig up in the future, and fz1p.c performs leaps and bounds ahead of it's programmatic predecessor.